system: implement a root overlay
This implements a overlayfs to create a writable root filesystem.

The overlay is backed by real storage after a disk is mounted to
/storage/.overlay/. This mountpoint is then bound to /overlay for
direct access.

The overlay is enabled by passing "rootoverlay" at the kernel
commandline.

/flash and /storage are mounted (again) atop the overlay so they
behave as normal.

rootoverlay is disabled when running off network storage. NFS is
unsuitable for an overlay's lower per kernel docs. Overlays may be
exported since kernel 4.16 so implement this on the server side if
desired. Other network mounting options are unknown so disabled.

Signed-off-by: Ian Leonard <antonlacon@gmail.com>