busybox: init: check unallocated space before setting resize marker
This check is to prevent a boot loop should someone set
SYSTEM_SIZE + STORAGE_SIZE to equal their memory device, or
/storage could not otherwise expand.

Determine parent node by taking STORAGE_TARGET, and removing the last
character. It then checks if the new string is a valid block device.
It continues to do so until the string is reduced to "/dev/", in the
event that a disk has many partitions.

A GPT disk could have up to 128 partitions (possibly more) and breaking
the loop early may yeild a false positive. /dev/sda100 would also trigger
on /dev/sda10 and /dev/sda1 if broken early, while we want /dev/sda.

We're interested in the unallocated space at the end of the device, so
be sure to take the last "Free Space" line output from parted when
parsing its output.

Signed-off-by: Ian Leonard <antonlacon@gmail.com>